Debugger: Some adjustments to the remote protocols
* WriteMemory: bytesWritten should not be size_t as it is architecture
  specific.
* RemoteDebuggerInterface::WriteMemory(): limit bytesWritten more
  consequently.
* Add management request DebugTeam.

+++ headers/private/debugger/debugger_interface/remote/RemoteDebugRequestDefs.h
@@ -189,7 +189,7 @@ DEFINE_REQUEST_AND_RESPONSE_STRUCTS(
 	),
 	(
 		status_t, 		error,
-		size_t,			bytesWritten
+		uint64,			bytesWritten
 	)
 )
